requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Env settings
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Auth / Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
email-validator==2.1.0.post1

# Alembic for migrations
alembic==1.13.0
